backend/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Rita Lopes <rita_lino@outlook.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.111.0"
uvicorn = "^0.30.1"
websockets = "^12.0"
pytest = "^8.2.2"
sqlalchemy = "^1.4.22"
pydantic = {extras = ["email"], version = "^2.7.4"}
psycopg2-binary = "^2.9.9"
pandas = "^2.2.2"
python-dotenv = "^1.0.1"
alembic = "^1.13.2"
pydantic-settings = "^2.3.4"
pyjwt = "^2.8.0"
pwdlib = {extras = ["argon2"], version = "^0.2.0"}
python-multipart = "^0.0.9"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.2"
pytest-cov = "^5.0.0"
taskipy = "^1.13.0"
ruff = "^0.5.0"
httpx = "^0.27.0"

[tool.pytest.ini_options]
pythonpath = "/backend"
addopts = '-p no:warnings'

[tool.taskipy.tasks]
lint = 'ruff check . && ruff check . --diff'
format = 'ruff check . --fix && ruff format .'
run = 'fastapi dev app/main.py'
lint_test = 'task lint'
test = 'pytest -s -x --cov=app -vv'
cov_test = 'coverage html'
run_local = 'fastapi dev app/main.py --host 0.0.0.0'

[tool.ruff]
line-length = 89
extend-exclude = ['migrations','alembic/versions' ,'venv', 'build', 'dist', 'htmlcov']

[tool.ruff.lint]
preview = true
select = ['I', 'F', 'E', 'W', 'PL', 'PT', 'C90', 'N' ]

[tool.ruff.format]
preview = true
quote-style = 'single'

[build-system]
requires = ["poetry-core"]
build-app = "poetry.core.masonry.api"
